+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0 (23.11.2021)
+
+* Customize the background color and the text color in the tooltips.
+* Center the tooltips.
+* Implement a Mutation Observer to create the tooltips.
+
 ## 1.0.0 (08.11.2021)
 
 * Added the respective tests.